requirements.txt
Flask==2.0.2
numpy==1.22.0
gunicorn==20.1.0
Pillow==9.0.1
requests==2.26.0
tensorflow-cpu==2.8.1
keras==2.8.0
Werkzeug==2.0.2
